Tidying up main.sml
We have accumulated quite a bit of technical debt in the argument-parsing
area; this really should be refactored out into a proper structure
and dealt with there. This is a much bigger job than I am willing to
take on right now, so it stays like this for a while longer.