Convert dictionaries to trees
This first big push converts dictionaries from the underlying “list”
representation to a “tree” representation. For most jobs there will be
no obvious difference. It will probably not run much faster as some
operations such as intersection or union still require the tree to be
converted to a list first. However operations such as insert and get
will be much faster.