Fast build from sorted list
Doing a naïve insert from a sorted list is very bad, building unbalanced
trees. Instead, we can exploit the fact it is sorted to build balanced
trees faster than just inserting each element!

We use this same method now to build regular trees, as there is probably
a good chance that users will supply sorted data. We sort the data just
to be sure, but at worst this is an additive of O(n) worse than before,
meaning it is still O(n), but we are guaranteed to have a more balanced
tree as a result. This is a net win.