Type- and file-based groundwork for importance
Set the types and file reading for importance calculations. We also
provide some default importances in the property_readers.sml file.
This commit also makes the first distinction between question properties
and representation properties in the code. Until now they were the
same thing, but now they split based on whether they carry importance
or not. To save some repetition we allow the use of an ‘intersection’
list of properties that have their importance stripped in representations.