Property and Correspondence
Introduce two new signatures (with three structures) for properties and
correspondences.

Properties are characterised by a comparison, and conversions to and
from strings. They are very simple. More importantly this file defines
sets and dictionaries over properties which the other modules can use.

Correspondences are more complicated, but this mostly captures many of
the common operations around them such as matching, equality, and if
they have the same properties.