Unfold and replicate
Introduce two new functions in robinlib.

`unfold` takes a generator function and a seed, and from it generates
a list. The generator function must eventually return `NONE`, else
unfold will never construct a complete list.

`replicate` takes a size and a value, and produces a list of the
given size filled entirely with that value.