Dictionaries update key value by function
While it is possible to update values with "insert", it is often
useful to specifically update a value based on the current value.
This would have required a get and an update as two separate calls:
the splay tree means this is fine, but in case the implementation
changes it is probably a good idea to make this an atomic operation.